+++ gatsby/src/components/contact/contact-menu-list.js
@@ -1,0 +1,36 @@
+import React from 'react';
+import {Link} from 'gatsby';
+
+const ContactMenuList = ({menu}) => {
+    console.log('contact menu')
+    console.log(menu);
+    return (
+        <>
+            {menu.length ? (
+                <>
+                    {menu.map((el, i) => (
+                        <div 
+                            className="column menu-item"
+                            key={i}
+                        >
+                            {el.object_slug !== "email" ? (
+                                <Link to={`${el.url}`} target="_blank">
+                                    {el.title}
+                                </Link>
+                            ) : (
+                                <a href={el.url}>{el.title}</a>
+                            )}
+                        </div>
+                    ))}
+                </>
+                ) : (
+                    <>
+                        Add Contact Menu items
+                    </>
+                )
+            }
+        </>
+    );
+}
+
+export default ContactMenuList;