+++ gatsby/src/components/ui/close.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const Close = () => {
+    return (
+        <div className={`close`}>
+            <div></div>
+            <div></div>
+        </div>
+    )
+}
+
+export default Close;