+++ gatsby/src/components/ui/contact-button.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import Close from './close';
+import ContactText from './contact-text';
+
+const ContactButton = ({expanded, onClick}) => {
+    return (
+        <div 
+            className={`contact-button ${expanded ? 'open' : ''}`}
+            onClick={onClick}>
+            <ContactText/>
+            <Close/>
+        </div>
+    );
+};
+
+export default ContactButton;